Remove visual_line_to_data from MIPS and RISCV modules
