+++ src/emulation_core.rs
@@ -3,3 +3,4 @@
 pub mod architectures;
 pub mod datapath;
 pub mod mips;
+pub mod line_info;
